project_root/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]==1.7.4
python-dotenv
python-multipart
pydantic[email]
